refactor(google): extract filterActions into separate utility function
see https://github.com/adobe/helix-deploy/pull/250#pullrequestreview-654014535
